fix: rename network error logging link
