+++ src/api/apollo.js
@@ -4,7 +4,7 @@ import BasketLinkCreator from './BasketLink';
 import ContentfulPreviewLink from './ContentfulPreviewLink';
 import ExperimentIdLink from './ExperimentIdLink';
 import HttpLinkCreator from './HttpLink';
-import GraphqlErrorLink from './GraphqlErrorLink';
+import NetworkErrorLoggingLink from './NetworkErrorLoggingLink';
 import NetworkErrorRetryLink from './NetworkErrorRetryLink';
 import SnowplowSessionLink from './SnowplowSessionLink';
 import { initState, setLocalState } from './localState';
@@ -40,7 +40,7 @@ export default function createApolloClient({
 	const client = new ApolloClient({
 		link: ApolloLink.from([
 			NetworkErrorRetryLink(),
-			GraphqlErrorLink(),
+			NetworkErrorLoggingLink(),
 			SnowplowSessionLink({ cookieStore }),
 			ExperimentIdLink(),
 			Auth0LinkCreator({ cookieStore, kvAuth0 }),
